Move EpochRewardsHasher to solana-sdk
